+++ Project_Spring/src/main/java/com/spring/project/service/ConfirmationTokenService.java
@@ -10,9 +10,9 @@ public interface ConfirmationTokenService {
 
     Optional<ConfirmationToken> getToken(String token);
 
-    void deleteByclient_Id(Integer id);
+    void deleteByclient_Id(Long id);
 
     void setConfirmedAt(String token);
 
-    ConfirmationToken findTokenByUserId(int id);
+    ConfirmationToken findTokenByUserId(Long id);
 }
